add `register` endpoint
